Revert "Generalized Connection Retrival"
This reverts commit c9a0514c5b50f3d4031250e31f488e120237eaf6.
